+++ src/base/hip/hip_matrix_mcsr.cpp
@@ -89,7 +89,7 @@ namespace rocalution
     }
 
     template <typename ValueType>
-    void HIPAcceleratorMatrixMCSR<ValueType>::AllocateMCSR(int nnz, int nrow, int ncol)
+    void HIPAcceleratorMatrixMCSR<ValueType>::AllocateMCSR(int64_t nnz, int nrow, int ncol)
     {
         assert(nnz >= 0);
         assert(ncol >= 0);
@@ -118,7 +118,7 @@ namespace rocalution
 
     template <typename ValueType>
     void HIPAcceleratorMatrixMCSR<ValueType>::SetDataPtrMCSR(
-        int** row_offset, int** col, ValueType** val, int nnz, int nrow, int ncol)
+        int** row_offset, int** col, ValueType** val, int64_t nnz, int nrow, int ncol)
     {
         assert(*row_offset != NULL);
         assert(*col != NULL);
