+++ src/base/host/host_matrix_bcsr.hpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2020-2022 Advanced Micro Devices, Inc. All rights Reserved.
+ * Copyright (C) 2020-2023 Advanced Micro Devices, Inc. All rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -51,11 +51,11 @@ namespace rocalution
         }
 
         virtual void Clear(void);
-        virtual void AllocateBCSR(int nnzb, int nrowb, int ncolb, int blockdim);
+        virtual void AllocateBCSR(int64_t nnzb, int nrowb, int ncolb, int blockdim);
         virtual void SetDataPtrBCSR(int**       row_offset,
                                     int**       col,
                                     ValueType** val,
-                                    int         nnzb,
+                                    int64_t     nnzb,
                                     int         nrowb,
                                     int         ncolb,
                                     int         blockdim);
